# executor: optimize cursor read point get by reading through pessimistic lock cache