--- a/executor/point_get.go
+++ b/executor/point_get.go
@@ -237,17 +237,23 @@ func (e *PointGetExecutor) Next(ctx context.Context, req *chunk.Chunk) error {
 				return err
 			}
 
-			e.handleVal, err = e.get(ctx, e.idxKey)
-			if err != nil {
-				if !kv.ErrNotExist.Equal(err) {
+			readFromCache := !e.ctx.GetSessionVars().IsPessimisticReadConsistency() || len(e.handleVal) > 0
+
+			if !readFromCache {
+				e.handleVal, err = e.get(ctx, e.idxKey)
+				if err != nil {
+					if !kv.ErrNotExist.Equal(err) {
+						return err
+					}
+				}
+			} else {
+				// try lock the index key if isolation level is not read consistency
+				// also lock key if read consistency read a value
+				err = e.lockKeyIfNeeded(ctx, e.idxKey)
+				if err != nil {
 					return err
 				}
-			}
-
-			// try lock the index key if isolation level is not read consistency
-			// also lock key if read consistency read a value
-			if !e.ctx.GetSessionVars().IsPessimisticReadConsistency() || len(e.handleVal) > 0 {
-				err = e.lockKeyIfNeeded(ctx, e.idxKey)
+				e.handleVal, err = e.get(ctx, e.idxKey)
 				if err != nil {
 					return err
 				}
@@ -377,7 +383,7 @@ func (e *PointGetExecutor) lockKeyIfNeeded(ctx context.Context, key []byte) erro
 			return err
 		}
 		lockCtx.IterateValuesNotLocked(func(k, v []byte) {
-			seVars.TxnCtx.SetPessimisticLockCache(kv.Key(k), v)
+			seVars.TxnCtx.SetPessimisticLockCache(k, v)
 		})
 		if len(e.handleVal) > 0 {
 			seVars.TxnCtx.SetPessimisticLockCache(e.idxKey, e.handleVal)