<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #36162

Problem Summary:

### What is changed and how it works?

For the point-get must lock the secondary key, lock it before reading, then later TiDB can read from pessimistic lock cache, this reduces one kv-get RPC. 

I test it by an edited sysbench update-non-index(wrap it in pessimistic txns) with the following options:

- clustered index off
- 32 threads
- 10,000,000 table size
- 1 table

#### Before

- TPS: 3174
- Latency(avg): 10.08ms
- Latency(.95): 16.71ms

<img width="457" alt="Screen Shot 2022-07-12 at 10 11 01 PM" src="https://user-images.githubusercontent.com/9587680/178524923-e30308c6-5b6e-44a9-956b-9f2380d04714.png">

#### After

- TPS: 3509
- Latency(avg): 9.12ms
- Latency(.95): 15.27ms

<img width="461" alt="Screen Shot 2022-07-12 at 11 14 33 PM" src="https://user-images.githubusercontent.com/9587680/178524954-8fa75461-850d-4565-a798-7f2f7fbb786b.png">

There are no kv-get requests.

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
Optimize cursor read point get by reading through pessimistic lock cache.
```