@you06 2022-07-12
This optimization only works when the following conditions are hit:

- pessimistic transaction(do not include pessimistic auto commit)
- repeatable read
- non-clustered PK or unique index

Maybe we can push down the lock-if-exist mechanism(RC mode) into TiKV, then it's able to reduce one round RPC for RC, cc @TonsnakeLin.